backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.1
pydantic==2.6.4
pytest==8.1.1
SQLAlchemy==2.0.29
python-dotenv==1.0.1
requests==2.31.0
openai>=1.0.0
